+++ src/components/molecules/article/ArticleList/index.spec.tsx
@@ -1,9 +1,9 @@
 import { mockArticles } from "@mocks/data";
 import { render, screen } from "jest/test-utils";
 
-import ArticleList from "../ArticleList";
+import ArticleList from ".";
 
-describe("components/molecules/ArticleList", () => {
+describe("components/molecules/article/ArticleList", () => {
   const mockArticleList = Object.values(mockArticles);
 
   it("OK: 表示が正しい", () => {
