+++ src/components/molecules/article/ArticleTipList/index.stories.tsx
@@ -4,7 +4,7 @@ import type { ComponentMeta, ComponentStory } from "@storybook/react";
 import { ArticleTipWithThumb } from ".";
 
 export default {
-  title: "molecules/ArticleTipWithThumb",
+  title: "molecules/article/ArticleTipWithThumb",
   component: ArticleTipWithThumb,
 } as ComponentMeta<typeof ArticleTipWithThumb>;
